pyproject.toml
[tool.poetry]
name = "apollo"
version = "2024.9"
description = "A robust data collection and analysis tool to document political processes"
authors = ["Tim Akinbo <takinbo@timbaobjects.com>"]
license = "GPL-3.0-only"
readme = "README.md"
repository = "https://github.com/nditech/apollo"
keywords = ["elections", "political-parties", "election-quick-count", "pvt", "election-analysis",  "election-data", "elections-toolkit"]

[tool.poetry.dependencies]
python = "~3.11.10"
aniso8601 = "0.82"
Arpeggio = "1.7.1"
awesome-slugify = "1.6.5"
boto3 = "1.16.8"
cachetools = "^5.5.0"
celery = "^5.4.0"
filedepot = "0.8.0"
Flask = "^3.0.3"
Flask-Admin = "^1.6.1"
flask-apispec = "^0.11.4"
Flask-Assets = "^2.1.0"
Flask-Cors = "^5.0.0"
Flask-Gravatar = "0.5.0"
Flask-HTTPAuth = "2.7.0"
Flask-JWT-Extended = "^4.6.0"
Flask-Mail = "0.9.1"
flask-menu = "0.7.0"
Flask-Migrate = "2.1.1"
Flask-Principal = "0.4.0"
flask-redis = "0.4.0"
Flask-Script = "2.0.6"
Flask-Security = "^5.5.2"
Flask-SQLAlchemy = "2.4.0"
Flask-SSLify = "0.1.4"
Flask-Uploads = "0.2.1"
Flask-Webpack = "0.1.0"
Flask-WTF = "^1.2.1"
GeoAlchemy2 = "0.6.3"
geojson = "2.5.0"
gevent = "^24.2.1"
GitPython = "^3.1.43"
gunicorn = "19.7.1"
honcho = "1.0.1"
loginpass = "0.3"
lxml = "^5.3.0"
Mako = "1.0.7"
marshmallow-sqlalchemy = "^0.27.0"
netifaces = "0.10.9"
networkx = "2.1"
numpy = "^2.1.1"
openpyxl = "2.5.0"
pandas = "^2.2.2"
parsimonious = "0.8.0"
Parsley = "1.3"
pillow = "^10.4.0"
prettyconf = "1.2.3"
prometheus-client = "0.6.0"
psutil = "5.6.6"
psycopg2-binary = "^2.9.9"
python-dateutil = "^2.9.0.post0"
python-editor = "1.0.3"
python-magic = "0.4.15"
pytz = "^2024.2"
qrcode = "6.1"
raven = "6.6.0"
redis = "3.2.1"
regex = "2018.2.3"
requests = "^2.32.3"
shapely = "^2.0.6"
six = "1.11.0"
SQLAlchemy = "^1.3.4"
sqlalchemy-json = "0.2.3"
SQLAlchemy-Utils = "^0.39.0"
tablib = "0.13.0"
Unidecode = "0.4.21"
whitenoise = "3.3.1"
wrapt = "1.12.1"
WTForms = "^3.1.2"
WTForms-Alchemy = "0.16.6"
xlrd = "1.1.0"
xlwt = "1.3.0"
greenlet = "^3.1.0"
pyxform = "~=0.11.1"
Authlib = "0.11"
cryptography = "3.3.2"
apispec = "^4.0.0"
flask-babel = "^4.0.0"

[tool.poetry.dev-dependencies]
autopep8 = "~=1.4.4"
coverage = "~=4.5.3"
flake8 = "~=3.7.7"
Flask-DebugToolbar = "0.10.1"
Flask-Testing = "0.8.1"
mimesis = "~=3.2.0"
pip-tools = "~=4.1.0"
pytest = "7.0.1"
selenium = "~=3.141.0"
chromedriver-autoinstaller = "~=0.6.2"
pytest-flask = "1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
